reworked fuzzer to use native generator
