+++ statetests.ini
@@ -3,7 +3,8 @@
 fork_config = Byzantium
 clients = geth
 
-tests_path = /tmp/testfiles
+artefacts = ~/tmp/evmlab/artefacts/
+tests_path = /tmp/evmlabtemp/
 random_tests = Yes
 prestate_tmp_file = prestate.json
 single_test_tmp_file = single_test_tmp.json
@@ -34,12 +35,3 @@ parity.docker_name = holiman/parity
 
 #testeth.docker_name = cdetrio/testeth
 
-[devops]
-
-# The fuzzer server uses raw binaries
-# instead of docker images
-
-cpp.binary         = /datadrive/evmlab/containers/bins/testeth
-parity.binary      = /datadrive/evmlab/containers/bins/parity-evm
-geth.binary        = /datadrive/evmlab/containers/bins/evm
-testeth.binary     = /datadrive/evmlab/containers/bins/testeth
